requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.12
sqlalchemy==2.0.35
alembic==1.13.3
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
itsdangerous==2.2.0
starlette==0.38.6
python-dotenv==1.0.1
